+++ server.h
@@ -15,5 +15,8 @@ void server_init (struct audio_s *audio_config_p, struct misc_config_s *misc_con
 
 void server_send_rec_packet (int chan, packet_t pp, unsigned char *fbuf,  int flen);
 
+int server_callsign_registered_by_client (char *callsign);
+
+
 
 /* end server.h */